fix(HeaderBar): dropdown fix
