+++ .changeset/purple-foxes-sing.md
@@ -1,0 +1,5 @@
+---
+'@talend/design-system': patch
+---
+
+HeaderBar isn't ready for primetime yet. API and style need work.
